+++ whisperlib/rpc/server_base.cc
@@ -1,6 +1,33 @@
 // Copyright: Urban Engines, Inc. 2012 onwards.
 // All rights reserved.
 // cp@urbanengines.com
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are
+// met:
+//
+// * Redistributions of source code must retain the above copyright
+// notice, this list of conditions and the following disclaimer.
+// * Redistributions in binary form must reproduce the above
+// copyright notice, this list of conditions and the following disclaimer
+// in the documentation and/or other materials provided with the
+// distribution.
+// * Neither the name of Urban Engines inc nor the names of its
+// contributors may be used to endorse or promote products derived from
+// this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
 
 #include "whisperlib/rpc/server_base.h"
 #include "whisperlib/rpc/rpc_http_server.h"
@@ -67,7 +94,7 @@ void ServerBase::Stop() {
 
 
 void ServerBase::StopInSelectThread() {
-    for (int i = 0; i < working_threads_.size(); ++i) {
+    for (size_t i = 0; i < working_threads_.size(); ++i) {
         working_threads_[i]->CleanAndCloseAll();
     }
     if (http_server_ != NULL) {
@@ -85,7 +112,7 @@ void ServerBase::Shutdown() {
     rpc_server_ = NULL;
     delete http_server_;
     http_server_ = NULL;
-    for (int i = 0; i < working_threads_.size(); ++i) {
+    for (size_t i = 0; i < working_threads_.size(); ++i) {
         delete  working_threads_[i];
     }
     working_threads_.clear();
@@ -108,7 +135,7 @@ void ServerBase::ToDebugHtmlTable(int64 id_val,
     http_req->request()->server_data()->Write(
         strutil::StringPrintf("<html><body><h1> Id lookup: %ld</h1>"
                               "<table border=1>", id_val));
-    for (int i = 0; i < data.size(); ++i) {
+    for (size_t i = 0; i < data.size(); ++i) {
         http_req->request()->server_data()->Write(
             strutil::StringPrintf("<tr><td>%s</td><td>%s</td></tr>\n",
                                   data[i].first.c_str(),
